Provide mount.btrfs helper
This patch provides a mount.btrfs helper for the mount command.
A btrfs filesystem could span several disks. This helper scans all the
partition to discover all the disks required to mount a filesystem.
So it is not necessary any-more to "scan" the partitions to mount a filesystem.

It adds in the option parameters the devices required to mount a filesystem.
Supposing that a filesystem is composed by several disks (/dev/sd[cdef]), when
the user runs "mount /dev/sdd /mnt", mount.btrfs is called and executes the
the mount(2) syscall as below:
mount("/dev/sdd", "/mnt", "btrfs", 0, "device=/dev/sdc,device=/dev/sde,device=/de/vsdf").

This helper uses both the libblkid and libmount to discover the devices, to
compute the parameters manipulation and to update the mtab file.

I got the idea from the btrfs.wiki; its biggest gains is to avoid the
separation of scanning phases (at boot time or during the block device
discovery) from the mounting. Also mkfs.btrfs could avoid to re-do a rescan of
the devices after a formatting.

mount.btrfs doesn't add more requirement than the mount command. It would be
possible to remove the "btrfs" command from the initramfs, and all the related
scripts (in my debian both udev and btrfs-tools packages contains three udev
rules for btrfs).

It is possible to set the timeout for waiting the device. After the timeout
is expired if the "degraded" options is used, it is tried to mount
the filesystem in degraded mode.
The parameter is called "device_timeout".